MOVE OF CONFIGURATION FILES INTO "DISTRIBUTED" ENVIRONMENT DIRECTORY
I want to repurpose this repo so that it can be used to test different
configurations corresponding to different environments, those
environments being as follows:

current = Simulation of current server topology that the Varilink
services are deployed to.

to-be = Simulation of planned, future server topology that the Varilink
services are to be deployed to. Relative to current this could reflect
Debian release upgrades, new hosts, etc.

distributed = Single service to single host deployment used on in this
Docker environment to confirm that deployment in this mode works.

What we've had so far in this repo represents the third of these
environments so this commit moves the associated configuraiton files
into a directory that defines this environment so that we can then
subsequently track changes to those files.