ADD MY-ROLES DIRECTORY
This contains a set of roles that are specific to the container
environments supported by this repository. These import the roles in
the libraries-ansible repository, replacing their handlers with ones
that work in containers where there is no service manager. The roles in
the libraries-ansible repository have been restructured to enable this
kind of selective import.