MOVE DOMAIN CONFIGS INTO DISTRIBUTED ENVS PROJECT PLAYBOOK GROUP_VARS
We were importing configuration files into playbooks that configured
domains. It's better to actually use the group_vars folder relative to
project specific playbooks. We'll move them to the distributed
environment as that was effectively the environment we were in when we
were using the domain configuration files. We can then track change
from that point.