MOVE DOCKER CONFIG OF THE DISTRIBUTED ENVIRONMENT INTO NEW STRUCTURE
The new structure is the one that I have adopted while testing using the
live environment. There will be changes to the distributed environment's
Docker configuration for sure, but I'll just move the files for now so
that changes to them can then be tracked.

Where there is a split between "deploy" and "test" aspects I've moved
the deploy aspect and deleted the test aspect. We have simplied our
approach so that we no longer have those two separate aspects.