ENABLE PASSTHROUGH OF DOCKER COMPOSE OPTIONS IN RAISE-HOSTS SERVICE
Now any arguments to the raise-hosts Docker Compose service that start
'-' are passed directly through the last docker-compose command that is
output, which is the one that brings the host emulation services up. The
most important use case for this is that it allows us to pass through
the --build option and so force a rebuild of existing images if
required.