COMMIT SERVICES CREATE PLAYBOOK IN THE LIVE ENV AFTER VERY BASIC TESTS
The "very basic tests" were to run the playbook for each of the services
and then undertake checks as identified below on a service-by-service
basis.

Backup:
Connect to the director using the console client and check the status of
the director, storage daemon and one of the clients that is up when the
scope is limited to backup.

Calendar:
Connect to radicale using the cadaver console client.

DNS:
Using the dig client accessing the internal DNS service lookup both the
host of the internal DNS service and some true, external domain. The
latter test just checks the upstream resolution path through to an
external DNS provider via the dns-external host.

Dynamic DNS:
Just confirm a clean run of the Ansible playbook. Without Linode
integration enabled this service does nothing functionally that could be
tested.

Mail:
Just confirm a clean run of the Ansible playbook. Until the project
playbooks are run there are no domain mail services to test.

Web:
Just confirm a clean run of the Ansible playbook. Until the project
playbooks are run there are no websites to test.