START BACULA DIRECTOR PROCESS WITHIN BASH SHELL
I have moved to starting all processes within the start task lists of
my-roles' roles within a bash shell. The only one I'd missed was the
start of the bacula-dir process within the backup_director role, so
correcting that.

This seems to help avoid processes dying, even though they're launched
with nohup, when the Ansible shell that started them closes, though I've
not definitively proved or explained that.