MOVE "LIVE" ENVIRONMENT FOLDER TO BE "NOW" ENVIRONMENT FOLDER
It's consistently awkward to differentiate between the test live
environment and the true live environment, both in READMEs and in blog
posts. So, I will rename the test live environment to "now", which sits
nicely alongside the "to-be" test environment.

This commit is just the moving of files to sit under envs/now rather
than envs/live so that we can track changes from that point.