REARRANGE PAUSE WHEN STARTING PROCESSES IN ROLES IN THE MY ROLES WRAPPER
I am still getting intermitent issues with processes failing to start. I
think it's because the task that start the process with nohup is exiting
too quickly for the process to get up and running before it terminates
the shell it is using.

So, I've put a sleep for three seconds immediately after the process
start in the same task. That way the pause happens before the task
completes. As a consequence, I've moved the sleep in the loop within the
"check" task that follows the "start" task