ENABLE DIG QUERIES IN NEW SPLIT DOCKER NETWORKS
Since I implemented the dig client I've split the Docker network into
the "office" and "internet" Docker networks and put in place routing
between the two. Now we must connect the dig client to one of the
networks, it doesn't matter which one since our dig wrapper explicitly
targets a DNS server. It's natural to connect it to the "office" Docker
network, which means we must also configure it so that it can route to
the "internet" Docker network when it sends queries to the external DNS
server.