MAKE ROUTER SERVICE WORK FOR DIFFERENT TEST ENVIRONMENTS
The router service as it was implemented was hard coded so that it would
only work with the "now" and "to-be" test environments. It wouldn't work
with the distributed test environment because the services within that
environment correspond to different IP addresses.

The router service now adapts to all three test environments. Note that
this is the reason why we changed dns.env files to network.env, since
this change requires that network details are added to thsoe files that
are not DNS related; see commit:
91fe3c33910012784ad51200904db71acddbf481