RENAME CUSTOMER PLAYBOOK FILE TO CREATE
I have implemented the means to have multiple playbook files within the
customer folder. There was a single file called playbook, which created
the customer services. Now that I can have multiple playbook files I'm
going to have one to create the customer serivces, one to delete them,
etc. So the first step is to rename the playbook file to create, since
this is what it does.