RENAME OF HOSTS IN NOW/TO-BE ENVS AND HOSTS.INI RESTRUCTURE THROUGHOUT
I've simplified some of the host names in the now and to-be test
environments and I've restructured the hosts.ini files in all three
test environments to give more groups that can be used in playbooks.

In this commit in respect of the Ansible artefacts it's only the
hosts.ini files that have changed along with renaming of affected
host_vars files. Obviously this also impacts playbooks but there are
other changes to be made to playbooks too so I've left them out of this
commit.