REMOVE HANDLERS FROM NONE ON IMPORTS IN MY ROLES WRAPPER
I have renamed the main.yml handlers files in the libraries-ansible
submodule to service.yml. This means that they won't now be auto
included when importing libraries-ansible roles within roles in my-roles
in this repository. Consequently, we no longer need to import with
handlers_from set to none.yml in roles in my-roles in this repository.
Indeed the none.yml handlers files have been removed in
libraries-ansible, so they're no longer available to be imported from
there.