ADD CONTAINER CONNECTED TO INTERNET NETWORK IN BACKUP GROUP
The backup group of containers brought up by the raise-hosts command
only included dns-external that is connected to the Docker internet
network. However, this host is excluded from backup, so we had no
means to test the backup service for a backup client on that network. I
have added the mail-external host to the backup group for all test
environments to resolve this.