CONNECT SEVERAL TESTING CLIENTS TO THE OFFICE NETWORK
We use to use the default network but now we don't. When we did those
services that didn't have "networks" configured would automatically
connect to the default network. Now we have to explicitly connect them
to the office network, which has effectively replaced the default
network in this Docker Compose configuration.