ADD VARILINK "LIBRARIES - ANSIBLE PLAYBOOKS" REPOSITORY AS A SUBMODULE
I have been using playbooks in this repository that are specific to the
Docker containers that it implements. I'm going to move to using a
shared library of playbooks across both Docker and Ansible managed
environments. Being shared, these playbooks are now in their own
repository so that it can be included as a submodule wherever those
playbooks are used.