REMOVE MY_ PREFIX FROM ALL ROLE NAMES IN THE MY-ROLES DIRECTORY
I've being using both the my-roles/ defined in this repository and the
roles that are included from the Varilink "Libraries - Ansible Roles"
repository as a submodule by having them both in the Ansible roles path.
This necessitated that those two sets of roles were separated by naming
them differently, hence the "my_" prefix on the role names in the
my_roles/ directory.

However, I've come to realize that it's cleaner to have my_roles/ in the
roles path and have them import the roles in "Libraries - Ansible Roles"
by specifying the complete path to those roles within containers. So,
the "my_" prefix on the roles in my_roles/ can be dropped. This is key
to being able to share the same playbooks to deploy to both Docker
containers and virtual servers as now we can use a common role name for
both scenarios within those playbooks.