ADD NETWORK CONFIGURATION FOR MONITOR GATEWAY EXTERNAL INTERFACE
The recently added monitor service is hosted on the Docker office
network but exposes a port to the Docker internet network. In the
"Services - Docker" environments I handle this as a proxy pass-through
configured within the router Docker Compose service. This change
contains that proxy configuration and also the setting of an environment
variable to dictate which host the traffic should be passed through to
for the distributed and now test environments. It seems that I had
already configured the environment variable in a previous commit
during testing of the to-be test environment.